[TDST-4] - Edited the following
Used only 1 h1 tag and replace the other h1 tags with h2 tags and h2 tags with h3 tags
Made "mark as done" icon color gray when it is not clickable
Reduced the font size for the body to 13px;
Edited the line chart y axis scaling similar to the angular material page.
Added more gap between the main container and the footer.